Fix the map2d get call for renderables not being offset

+++ core/src/main/scala/mwittmann/spooktober/screen/View.scala
@@ -1,14 +1,16 @@
 package mwittmann.spooktober.screen
 
+// Todo: At the moment, 'gameX, gameY' are the center of the screen, with the screen going from -(screenWidth/2, ..)
+// to +(screenWidth/2, ...); might be nicer to set it to have gameX, gameY be lower right
 class View(val gameX: Float, val gameY: Float, val gameWidth: Float, val gameHeight: Float, val screenX: Float, val screenY: Float, val screenWidth: Float, val screenHeight: Float) {
   val xFactor: Float  = screenWidth / gameWidth
   val yFactor: Float = screenHeight / gameHeight
 
-  def translateX(x: Float): Int = ((x - gameX) * xFactor + screenX).toInt
+  def translateX(x: Float): Float = (x - gameX) * xFactor + screenX
 
-  def translateY(y: Float): Int = ((y - gameY) * yFactor + screenY).toInt
+  def translateY(y: Float): Float = (y - gameY) * yFactor + screenY
 
-  def translateWidth(x: Float): Int = (x * xFactor).toInt
+  def translateWidth(x: Float): Float = x * xFactor
 
-  def translateHeight(y: Float): Int = (y * yFactor).toInt
+  def translateHeight(y: Float): Float = y * yFactor
 }
